requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
python-dotenv==1.0.1
pyserial==3.5
websockets==12.0
pydantic==2.6.4
python-multipart==0.0.9
openai==1.14.3
aiofiles==23.2.1
deep-translator==1.11.4
